HadoopUtils and YarnUtils improvements
