+++ streamx-console/streamx-console-service/src/test/java/SendEmailTest.java
@@ -18,8 +18,8 @@
  */
 
 import com.streamxhub.streamx.common.util.DateUtils;
-import com.streamxhub.streamx.common.util.HadoopUtils;
 import com.streamxhub.streamx.common.util.Utils;
+import com.streamxhub.streamx.common.util.YarnUtils;
 import com.streamxhub.streamx.console.core.entity.Application;
 import com.streamxhub.streamx.console.core.entity.SenderEmail;
 import com.streamxhub.streamx.console.core.enums.FlinkAppState;
@@ -131,7 +131,7 @@ public class SendEmailTest {
         }
         duration = duration / 1000 / 60;
         String format = "%s/proxy/%s/";
-        String url = String.format(format, HadoopUtils.getRMWebAppURL(false), application.getAppId());
+        String url = String.format(format, YarnUtils.getRMWebAppURL(false), application.getAppId());
 
         MailTemplate template = new MailTemplate();
         template.setJobName(application.getJobName());
