+++ streamx-flink/streamx-flink-shims/pom.xml
@@ -22,7 +22,7 @@
     <profiles>
 
         <profile>
-            <id>shims_2.12</id>
+            <id>flink-shims_2.12</id>
             <activation>
                 <activeByDefault>true</activeByDefault>
             </activation>
@@ -33,7 +33,7 @@
         </profile>
 
         <profile>
-            <id>shims_2.11</id>
+            <id>flink-shims_2.11</id>
             <activation>
                 <property>
                     <name>scala.binary.version</name>
